Set adbkey on all builds, not just during KWE command
It looks like running `adb connect ip:5555` from another machine still requires the `adbkey` to be set - i.e. the ADB key is not just used for USB debugging, but authorization over Wi-Fi as well. Just make all build and developer machines use the same key to avoid the authorization nonsense. We'd prefer to put our key at a unique path for UET and pass the key path via an environment variable or `adb connect`, but ADB doesn't support this so we just have to override the key globally.
